requirements.txt
pandas
numpy
sqlite
unittest
mysql-connector-python